deliverable-2/lexata-code/data-scripts/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.0
et-xmlfile==1.1.0
fonttools==4.29.1
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
openai==0.14.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandas-stubs==1.2.0.49
Pillow==9.0.1
plotly==5.6.0
pymongo==4.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
sec-api==1.0.9
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
tenacity==8.0.1
threadpoolctl==3.1.0
tqdm==4.62.3
urllib3==1.26.8
